Move env to root src, add skip validation in dockerfile build as the validation does not result in proper environment variables in runtime.
